+++ consensus/e2e_tests/utils_test.go
@@ -3,7 +3,6 @@ package e2e_tests
 import (
 	"context"
 	"fmt"
-	"log"
 	"os"
 	"reflect"
 	"sort"
@@ -505,24 +504,15 @@ func baseStateMachineMock(t *testing.T, _ modules.EventsChannel, bus modules.Bus
 	stateMachineMock.EXPECT().SetBus(gomock.Any()).Return().AnyTimes()
 	stateMachineMock.EXPECT().GetModuleName().Return(modules.StateMachineModuleName).AnyTimes()
 	//stateMachineMock.EXPECT().SendEvent(gomock.Any()).Return(nil).AnyTimes()
-	stateMachineMock.EXPECT().SendEvent(gomock.Any()).DoAndReturn(func(event coreTypes.StateMachineEvent, args ...any) error {
-		switch event {
-		case coreTypes.StateMachineEvent_Consensus_IsEnableServer:
-			log.Println("Enable Server mode")
-			bus.GetConsensusModule().EnableServerMode()
-		case coreTypes.StateMachineEvent_Consensus_IsDisableServer:
-			bus.GetConsensusModule().DisableServerMode()
-		//case coreTypes.StateMachineEvent_Consensus_IsUnsynched:
-		//
+	// stateMachineMock.EXPECT().SendEvent(gomock.Any()).DoAndReturn(func(event coreTypes.StateMachineEvent, args ...any) error {
+	// 	switch event {
 
-		//case coreTypes.StateMachineEvent_Consensus_IsSyncing:
+	// 	default:
+	// 		log.Println("asdads")
+	// 	}
+	// 	return nil
 
-		default:
-			log.Println("asdads")
-		}
-		return nil
-
-	}).AnyTimes()
+	// }).AnyTimes()
 
 	return stateMachineMock
 }