+++ shared/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.35] - 2023-02-27
+
+- Add FSMConsensusEvents() interface to handle state transition events in consensus module
+
 ## [0.0.0.34] - 2023-02-24
 
 - Remove SetLeaderId() method from ConsensusDebugModule interface